requirements.txt
anthropic>=0.40.0
streamlit>=1.40.0
tavily-python>=0.5.0
chromadb>=0.5.0
sentence-transformers>=3.0.0
requests>=2.32.0
beautifulsoup4>=4.12.0
pypdf2>=3.0.0
arxiv>=2.1.0
python-dotenv>=1.0.0
